+++ analog/main.py
@@ -47,61 +47,52 @@ def main(argv=None):
     parser.add_argument('--version',
                         action='version',
                         version="analog {v}".format(v=analog.__version__))
-    # either -f / --format or -r / --regex for format name or expression
-    formatargs = parser.add_mutually_exclusive_group()
-    # -f / --format
-    formatargs.add_argument('-f', '--format',
-                            action='store',
-                            dest='format',
-                            default=None,
-                            choices=format_choices,
-                            help="Log format")
-    # -r / --regex
-    formatargs.add_argument('-r', '--regex',
-                            action='store',
-                            help='regex format pattern with named groups.')
+
+    # common arguments
+    common = argparse.ArgumentParser(add_help=False)
+
     # -o / --output_format
-    parser.add_argument('-o', '--output-format',
+    common.add_argument('-o', '--output-format',
                         action='store',
                         dest='output_format',
                         default='plain',
                         choices=output_choices,
                         help="output format")
     # -p / --path
-    parser.add_argument('-p', '--path',
+    common.add_argument('-p', '--path',
                         action='append',
                         dest='paths',
                         default=DEFAULT_PATHS,
                         help="paths to monitor (repeat for multiple)")
     # -v / --verb
-    parser.add_argument('-v', '--verb',
+    common.add_argument('-v', '--verb',
                         action='append',
                         dest='verbs',
                         default=DEFAULT_VERBS,
                         help="verbs to monitor (repeat for multiple)")
     # -s / --status
-    parser.add_argument('-s', '--status',
+    common.add_argument('-s', '--status',
                         action='append',
                         dest='status_codes',
                         default=DEFAULT_STATUS_CODES,
                         help="status codes to monitor (repeat for multiple)")
     # -a / --max_age
-    parser.add_argument('-a', '--max_age',
+    common.add_argument('-a', '--max_age',
                         action='store',
                         type=int,
                         default=None,
                         help="analyze logs until n minutes age")
     # -ps / --path_stats
-    parser.add_argument('-ps', '--path-stats',
+    common.add_argument('-ps', '--path-stats',
                         action='store_true',
                         dest='path_stats',
                         help="include statistics per path")
     # -t / --timing
-    parser.add_argument('-t', '--timing',
+    common.add_argument('-t', '--timing',
                         action='store_true',
                         help="print timing")
     # logfile, defaults to stdin
-    parser.add_argument('log',
+    common.add_argument('log',
                         action='store',
                         nargs='?',
                         type=argparse.FileType('r'),
@@ -109,21 +100,54 @@ def main(argv=None):
                         help="logfile to analyze."
                              "Defaults to stdin for piping.")
 
+    # subcommands for predefined log formats
+    format_parsers = parser.add_subparsers(
+        title="log format",
+        description="analyze logfiles of a certain format",
+        metavar='FORMAT',
+        dest='format')
+    for format in format_choices:
+        format_parsers.add_parser(format, parents=[common],
+                                  help="{} log format".format(format))
+
+    # subcommand for custom log format
+    custom_format = format_parsers.add_parser('custom', parents=[common],
+                                              help="custom log format")
+    # -pr / --pattern-regex
+    custom_format.add_argument('-pr', '--pattern-regex',
+                               action='store',
+                               dest='pattern',
+                               required=True,
+                               help='regex format pattern with named groups.')
+    # -tf / --time-format
+    custom_format.add_argument('-tf', '--time-format',
+                               action='store',
+                               dest='time_format',
+                               required=True,
+                               help='timestamp format (strftime compatible)')
+
     try:
         if argv is None:
             argv = sys.argv
         args = parser.parse_args(argv[1:])
 
+        format_kwargs = {'format': args.format}
+        if args.format == 'custom':
+            format_kwargs.update({
+                'pattern': args.pattern,
+                'time_format': args.time_format,
+            })
+
         # analyze logfile and generate report
         analog.analyze(log=args.log,
-                       format=args.format or args.regex,
                        paths=args.paths,
                        verbs=args.verbs,
                        status_codes=args.status_codes,
                        max_age=args.max_age,
                        path_stats=args.path_stats,
                        timing=args.timing,
-                       output_format=args.output_format)
+                       output_format=args.output_format,
+                       **format_kwargs)
 
         parser.exit(0)
 
@@ -132,36 +156,3 @@ def main(argv=None):
 
     except KeyboardInterrupt:
         parser.exit(1, "\nExecution cancelled.")
-
-
-def _option_invalid_choice(option, value, choices):
-    """Error message for invalid choices of ini config options.
-
-    :param option: name of configuration option
-    :type option: ``str``
-    :param value: user-provided configuration value
-    :param value: ``str``
-    :param choices: valid option value choices
-    :type choices: ``list``
-    :returns: error message
-    :rtype: ``str``
-
-    """
-    return ("config option {option}: invalid choice {value!r} (choose from "
-            "{choices})".format(option=option, value=value,
-                                choices=", ".join(map(repr, choices))))
-
-
-def _option_mutually_exclusive(option1, option2):
-    """Error message for mutually exclusive config options specified together.
-
-    :param option1: name of first option
-    :type option1: ``str``
-    :param option2: name of second option
-    :type option2: ``str``
-    :returns: error message
-    :rtype: ``str``
-
-    """
-    return ("config option {option2} not allowed with "
-            "option {option1}".format(option1=option1, option2=option2))